+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Unreleased
 - lib.images.base: Fix keyword usage, naming collision
 - lib.platforms.director.unlock_dir_file: Add helper for unlocking Director files
 - lib.games.lemmings: Replace BitReader with BitStream
+- fields: Annotate errors with position in the block tree
 
 0.7.3 - 2019-12-07
 ==================